requirements.txt
streamlit>=1.32.0
requests>=2.31.0
pdfplumber>=0.10.0
pymupdf>=1.23.0
python-docx>=1.0.0
beautifulsoup4>=4.12.0
selenium>=4.18.0
webdriver-manager>=4.0.0
feedparser>=6.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
lxml>=5.0.0
pandas>=2.0.0
